+++ preferences.py
@@ -9,38 +9,25 @@ from bpy.props import (FloatProperty,
                         PointerProperty)
 
 def get_addon_prefs():
-    '''
-    function to read current addon preferences properties
-    access with : get_addon_prefs().super_special_option
-    '''
-    return bpy.context.preferences[__package__].preferences
-
-    # import os
-    # addon_name = os.path.splitext(__name__)[0]
-    # preferences = bpy.context.preferences
-    # addon_prefs = preferences.addons[addon_name].preferences
-    # return (addon_prefs)
-
+    return bpy.context.preferences.addons[__package__].preferences
 
 class STORYTOOLS_prefs(bpy.types.AddonPreferences):
     bl_idname = __name__.split('.')[0] # or with: os.path.splitext(__name__)[0]
 
     # some_bool_prop to display in the addon pref
-    super_special_option : bpy.props.BoolProperty(
+    default_edit_line_opacity : bpy.props.FloatProperty(
         name='Use super special option',
-        description="This checkbox toggle the use of the super special options",
-        default=False)
+        description="Edit line opacity for newly created objects\
+            \nSome users prefer to set it to 0 (show only selected line in edit mode)\
+            \nBlender default is 0.5",
+        default=0.2, min=0.0, max=1.0)
 
     def draw(self, context):
             layout = self.layout
-
-            ## some 2.80 UI options
-            # layout.use_property_split = True
+            layout.use_property_split = True
             # flow = layout.grid_flow(row_major=True, columns=0, even_columns=True, even_rows=False, align=False)
             # layout = flow.column()
-
-            # TODO: 
-
+            layout.prop(self, 'default_edit_line_opacity')
 
 ### --- REGISTER ---